+++ NEWS.md
@@ -1,5 +1,9 @@
 # purrr 0.2.2.9000
 
+* `as_function()` now generates a better argument list so you can refer to
+  arguments by position like `..1` and `..2`. This makes it possible to use
+  the formula shorthand for functions that need more arguments (#289).
+
 * New `reduce2()` and `reduce2_right()` make it possible to reduce with a
   3 argument function where the first argument is the accumulated value, the
   second argument is `.x`, and the third argument is `.y` (#163).